add a setState to the DualIconButton and allow the ortho/persp state to be initialized via json loading
