+++ agave_app/ViewToolbar.cpp
@@ -1,5 +1,7 @@
 #include "ViewToolbar.h"
 
+#include "renderlib/CCamera.h"
+
 #include <QFile>
 #include <QHBoxLayout>
 #include <QIcon>
@@ -9,56 +11,6 @@
 #include <QPixmapCache>
 #include <QPushButton>
 #include <QSvgRenderer>
-
-class DualIconButton : public QPushButton
-{
-private:
-  QIcon icon1;
-  QIcon icon2;
-  QString tooltip1;
-  QString tooltip2;
-  QString statustip1;
-  QString statustip2;
-  int state;
-
-public:
-  DualIconButton(const QIcon& icon1,
-                 const QIcon& icon2,
-                 const QString& tooltip1,
-                 const QString& statustip1,
-                 const QString& tooltip2,
-                 const QString& statustip2,
-                 QWidget* parent = nullptr)
-    : icon1(icon1)
-    , icon2(icon2)
-    , tooltip1(tooltip1)
-    , tooltip2(tooltip2)
-    , statustip1(statustip1)
-    , statustip2(statustip2)
-    , state(0)
-    , QPushButton(parent)
-  {
-
-    setIcon(icon1);
-    setToolTip(tooltip1);
-    setStatusTip(statustip1);
-    connect(this, &QPushButton::clicked, this, &DualIconButton::toggleIcon);
-  }
-
-  void toggleIcon()
-  {
-    state = 1 - state;
-    if (state == 1) {
-      setIcon(icon2);
-      setToolTip(tooltip2);
-      setStatusTip(statustip2);
-    } else {
-      setIcon(icon1);
-      setToolTip(tooltip1);
-      setStatusTip(statustip1);
-    }
-  }
-};
 
 ViewToolbar::ViewToolbar(QWidget* parent)
   : QWidget(parent)
@@ -156,3 +108,9 @@ ViewToolbar::ViewToolbar(QWidget* parent)
 }
 
 ViewToolbar::~ViewToolbar() {}
+
+void
+ViewToolbar::initFromCamera(const CCamera& camera)
+{
+  orthoViewButton->setState((camera.m_Projection == ProjectionMode::ORTHOGRAPHIC) ? 1 : 0);
+}
